Do not indefinitively report the compilation result for kernelci.
Without this patch, a crontab will report again and again. In order to
tell the report script to not continue, we put a tag 'report' telling
the branch has already been reported.

All the other changes are to silent the script in order to have empty
message when nothing has to be done in order to prevent to have an
email with the crontab output.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>