pmwg-ci: Initialize correctly and get the baseline branch
This change group all setup functions into a bigger one which make
sure the underlying functions are called in the right order.

The baseline branch name is retrieved from the configuration file
in order to get the tag from the branch itself and not in a global
manner. That will allow to specify if we track <branch>/HEAD or
<branch>/TAG change.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>