pmwg-ci: Fix baseline used as remote change
With the current code, we consider the baseline as one regular remote but
that is not true as we have specific policy for modified and updated baseline
(tag or commit-id changes). That triggers multiple times an update of the CI
loop for the same commit-id as it is assumed we have a remote branch updated
when the baseline is changed (and we want to track per tag).

Fix this by discarding the baseline when updating the remotes.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>