ci-merge: Require REMOTE_REPO and RR_CACHE to be specified explicitly
REMOTE_REPO and RR_CACHE can be made optional, so remove their default values
and make code that depends on them conditional on their being defined. This
allows someone to to locally run ci-merge without any rerere cache and
without requiring to push anything to a remote repository.

This CHANGES the default behaviour of ci-merge to require the -r and -c
options for REMOTE_REPO and RR_CACHE respectively.

To make things more robust, add a new function to check if repos specified
on the cmdline exist. We end up with some strange errors because we don't
error out early enough currently.

Signed-off-by: Amit Kucheria <amit.kucheria@linaro.org>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>