ci-merge: enfore merge.conflictstyle is set to 'merge'
merge.conflictstyle can be either 'merge' or 'diff3'. When using 'diff3' there
is more context information provided by git for each merge conflict. While git
rerere will try to be smart to discard the additional data about 'common
ancestor', it is possible that the conflict left when making the same merge are
actually different when using these two conflict styles.

IOW, if the merge produces

    <<<
    side A
    |||
    common
    ===
    side B
    >>>

when diff3 style is chosen, but if the same merge results in

    <<<
    side A'
    ===
    side B'
    >>>

where side A' is not identical to side A (or B' and B are not
identical), then we will fail to find the previously recorded
resolution.

as discussed here: https://marc.info/?l=git&m=153003225927418&w=2

So basically, it's safer to force a specific merge conflcit style in automerge
to avoid this type of issues, especially when sharing a rr-cache within the
team.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
Acked-by: Amit Kucheria <amit.kucheria@linaro.org>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>