ci-merge: Allow running mergetool without X
Allow running mergetool in a non-windowing environment, by spawning
mergetool inline from the script. We no longer need to guess what terminal
might be installed on the system.

As the config file is fed line by line into the while loop we need to dup
the original stdin, or the spawned mergetool will read stdin from the
config file.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
[refactored to master]
Signed-off-by: Amit Kucheria <amit.kucheria@linaro.org>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>