ci-merge: properly filter out remotes from local tree
The purpose of this function is to remove remotes from the local trees
if they have been removed from the config file.

It is broken since 2ebf50e36955 (ci-merge: Add missing -q to grep when
searching for old branches). When we have -q everywhere like:

echo $LINE | grep -q "^$REMOTE_NAME\b" | grep -q $REMOTE_URL | grep -q "$REMOTE_BRANCH"

Then it will *always* be false, since the first grep command will
return an empty string.

What we really wanted to do is to find a remote in the local tree
which is *not* in the config, so that we can simply remove it. We
don't even need to look/match for branches, e.g. if we have a remote
with the same name and the same url, then it's found, and we can keep
it in case it's in the config file.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>