ci-merge: fix a problem of not merged remote into integration branch
If there are two remotes and one of their remote names matches another
one as a regexp pattern, then it will not be added as a new remote.
This condition is met, when one remote name is a substring of another,
and it may result in a missing merge of a remote branch.

The problem is encountered by me, but the fix is attributed to Dmitry
Baryshkov.

Signed-off-by: Vladimir Zapolskiy <vladimir.zapolskiy@linaro.org>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>